refactor(router): move urlParser back up to router folder
