+++ modules/angular2/src/router/rules/rule_set.ts
@@ -13,7 +13,7 @@ import {
 } from '../route_config/route_config_impl';
 import {AsyncRouteHandler} from './route_handlers/async_route_handler';
 import {SyncRouteHandler} from './route_handlers/sync_route_handler';
-import {Url} from './url_parser';
+import {Url} from '../url_parser';
 import {ComponentInstruction} from '../instruction';
