+++ modules/angular2/src/router/url_parser.ts
